Change UpdateTogether mixin to UpdateTogetherModel
This'll work, but it sucks for a number of reasons:
1) I'm replacing models.Model just to add one damn field to Meta
2) If anyone else needed to do the same thing (and I certainly have
places where I do this already in other projects) I've just screwed
them. You either need to cook up your own metaclass that inherits from
all of our metaclasses and then do the same for your models.Model or
only use one of our packages.

This is dumb.