Actually support ForeignKeys without spawning needless queries.
This one’s my bad for not recognizing that FK relations are hidden
behind a descriptor, so hasattr() type checks will *also* fire a
subquery.

To fix this, I’ve removed all those checks, cleaned up the handling for
relation fields (properly grabbing the object if it exists, otherwise
stashing the PK, and handling the difference when needed), and swapped
out the old_values dict for a readonly mapping that defers everything,
so we won’t spawn a query unless you really need the associated data.