Only check mutability on first access after init/save.
This should be safe and is a bit of a performance improvement.

Since we only care about storing the original (here meaning in db)
value of a field, we only need to guard against that original value
being tainted. So simply checking the first time the attribute is
accessed is just fine, and saves us from paying the cost of the
recursive is_mutable() check as much as possible.